refactor: foldering component files
