+++ src/pages/index.tsx
@@ -1,14 +1,14 @@
 import Image from 'next/image';
 import { useRouter } from 'next/router';
-import Button from '../components/button';
-import Contacts from '../components/contacts';
-import Container from '../components/container';
-import Footer from '../components/footer';
-import Navbar from '../components/navbar';
-import PageHead from '../components/page-head';
-import Parallax from '../components/parallax';
-import Projects from '../components/projects';
-import Wrapper from '../components/wrapper';
+import Button from '../components/Button';
+import Contacts from '../components/Contacts';
+import Container from '../components/Container';
+import Footer from '../components/Footer';
+import Navbar from '../components/Navbar';
+import PageHead from '../components/PageHead';
+import Parallax from '../components/Parallax';
+import Projects from '../components/Projects';
+import Wrapper from '../components/Wrapper';
 import projects from '../data/projects.json';
 import { Project as TProject } from '../types/Project';
 import translation from '../utils/translation';
